Feat: jpa 키타입 수정

+++ user/src/main/java/org/newshabit/app/user/infrastructure/adapter/outbound/persistence/repository/UserDailyGoalRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDailyGoalRepository extends JpaRepository<UserDailyGoalEntity, Long> {
+public interface UserDailyGoalRepository extends JpaRepository<UserDailyGoalEntity, Integer> {
 
 	@Query("""
 		SELECT dailyGoal FROM UserDailyGoalEntity dailyGoal
